+++ src/components/SectionMain/componentsMain/BlockEducation.jsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import s from '../SectionLeft.module.css'
+import s from '../SectionMain.module.css'
 
 const BlockEducation = ({store, active}) => {
+
     return (
         <div className={active ? `${s.content} ${s.active}` : s.content}>
-            <span>{store.description}</span><br/>
-            <span>{store.time} | {store.place} </span>
-            <h4 className={s.from_title}>{store.position}</h4>
+            <h4>{store.description}</h4>
+            <span className={s.from_title}>{store.position}</span><br/>
+            {store.time&&<span>{store.time}</span>}
         </div>
     );
 };
 export default BlockEducation;
+
+
