+++ src/components/SectionMain/componentsMain/BlockProject.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import s from "../SectionLeft.module.css";
+import s from "../SectionMain.module.css";
 
 const BlockProject = ({project, active}) => {
     let arrProject = project.list
